Move LDAP helpers
